// src/common/enums/error-code.enum.ts
export enum ErrorCode {
  // Not Found(404)
  CONCERT_NOT_FOUND = 'CONCERT_NOT_FOUND',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  ARTIST_NOT_FOUND = 'ARTIST_NOT_FOUND',
  SETLIST_NOT_FOUND = 'SETLIST_NOT_FOUND',
  COMMENT_NOT_FOUND = 'COMMENT_NOT_FOUND',
  SONG_NOT_FOUND = 'SONG_NOT_FOUND',
  SETLIST_SONG_NOT_FOUND = 'SETLIST_SONG_NOT_FOUND',
  SETLIST_CONCERT_NOT_FOUND = 'SELIST_CONCERT_NOT_FOUND',
  GENRE_NOT_FOUND = 'GENRE_NOT_FOUND',

  // Bad Request(400)
  INVALID_CURSOR_FORMAT = 'INVALID_CURSOR_FORMAT',
  INVALID_ID_FORMAT = 'INVALID_ID_FORMAT',
  SEARCH_KEYWORD_REQUIRED = 'SEARCH_KEYWORD_REQUIRED',
  USER_ALREADY_DELETED = 'USER_ALREADY_DELETED',
  NICKNAME_ALREADY_EXISTS = 'NICKNAME_ALREADY_EXISTS',
  USER_ALREADY_EXISTS = 'USER_ALREADY_EXISTS',

  // Forbidden(403)
  USER_DELETED = 'USER_DELETED',
  COMMENT_DELETE_FORBIDDEN = 'COMMENT_DELETE_FORBIDDEN',
  WITHDRAWAL_PERIOD_NOT_PASSED = 'WITHDRAWAL_PERIOD_NOT_PASSED',
  KAKAO_USER_INFO_FETCH_HEAD = 'KAKAO_USER_INFO_FETCH_HEAD',

  // Unauthorized(401)
  APPLE_KEY_NOT_FOUND = 'APPLE_KEY_NOT_FOUND',
  REFRESH_TOKEN_INVALID = 'REFRESH_TOKEN_INVALID',
  REFRESH_TOKEN_EXPIRED = 'REFRESH_TOKEN_EXPIRED',
  REFRESH_TOKEN_VERIFICATION_FAILED = 'REFRESH_TOKEN_VERIFICATION_FAILED',
  REFRESH_TOKEN_NOT_FOUND = 'REFRESH_TOKEN_NOT_FOUND',

  // Internal Server Error(500)
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
}